Add viewport meta tag for better mobile view
The viewport is requested to be a minimum of 480px wide, and to maintain
an initial scale of 1. For devices which are narrower than 480 display
pixels, this results in a horizontal scroll bar. For devices which are
wider, the width is treated a minimum, and the full device width will be used.

A few additional styling changes have been made to better support mobile:
 - <h1> heading have been reduced from 6em to 2.5em on narrow screens.
 - The log/output boxed are now 100% width, rather than 600px exactly.
 - Some CSS rules have been pulled up into the global file.

Additional information can be found in https://developer.mozilla.org/en-US/docs/Web/HTML/Viewport_meta_tag