Set up a progressive web app manifest
This will attempt to locally cache all the current tools for offline capabilities
in modern browser, and will enable installing the site as an 'app' on
compatible devices.

Any request not in the cache will be requested from server. Requests in the cache
will be given 1.5s to complete or error on the origin before the cached response
is used. If an updated upstream resource is received, it will update the cache.